Add reconnection window
Important: Dont use, program crashes if a query fails, because null pointer
is returned, but no treatament is done yet.

If a query fails null pointer is returned and the reconnection window is
showed.

The reconnection window is set as the active window if any query is tried after
disconnect.