Permit especial treat to expected errors on DB functions.
Add expectedErros parameter on DB functions, a list of expected SQL
states of error. This parameter is empty by default.

If an error is listed as expected, then safeBDExec fails silently and
throws an exception corresponding to this position on expectedErros.

Rename PatientBDModel::DBExec(string command) to DBExecCommand to avoid
ambiguity.