Add autosave.
Add autosave on Patient tabs and AppointmentWidget content. Save text 20
seconds after any user writing and before update on DB. The text is
saved in local files named by the time the edit text widget was created
and the number of the widget within the ones of this type in this
program execution. The first line contains the title of the saved text
to identify it.